+++ apps/ledger-live-desktop/src/renderer/screens/exchange/Swap2/Form/SwapWebView.tsx
@@ -24,10 +24,7 @@ import {
 import { useRedirectToSwapHistory } from "../utils/index";
 
 import { SwapLiveError } from "@ledgerhq/live-common/exchange/swap/types";
-import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
 import { LiveAppManifest } from "@ledgerhq/live-common/platform/types";
-import { Box, Button } from "@ledgerhq/react-ui";
-import { t } from "i18next";
 import { usePTXCustomHandlers } from "~/renderer/components/WebPTXPlayer/CustomHandlers";
 import { captureException } from "~/sentry/internal";
 
@@ -67,24 +64,12 @@ export type SwapWebProps = {
   liveAppUnavailable(): void;
   sourceCurrencyId?: string;
   targetCurrencyId?: string;
+  isMaxEnabled?: boolean;
 };
 
 export const SwapWebManifestIDs = {
   Demo0: "swap-live-app-demo-0",
   Demo1: "swap-live-app-demo-1",
-};
-
-export const useSwapLiveAppManifestID = () => {
-  const demo0 = useFeature("ptxSwapLiveAppDemoZero");
-  const demo1 = useFeature("ptxSwapLiveAppDemoOne");
-  switch (true) {
-    case demo1?.enabled:
-      return demo1?.params?.manifest_id ?? SwapWebManifestIDs.Demo1;
-    case demo0?.enabled:
-      return demo0?.params?.manifest_id ?? SwapWebManifestIDs.Demo0;
-    default:
-      return null;
-  }
 };
 
 const SwapWebAppWrapper = styled.div`
@@ -98,6 +83,7 @@ const SwapWebView = ({
   liveAppUnavailable,
   sourceCurrencyId,
   targetCurrencyId,
+  isMaxEnabled,
 }: SwapWebProps) => {
   const {
     colors: {
@@ -112,12 +98,11 @@ const SwapWebView = ({
   const locale = useSelector(languageSelector);
   const redirectToHistory = useRedirectToSwapHistory();
   const enablePlatformDevTools = useSelector(enablePlatformDevToolsSelector);
-  const manifestID = useSwapLiveAppManifestID();
-  const isDemo1Enabled = manifestID?.startsWith(SwapWebManifestIDs.Demo1);
+
   const hasSwapState = !!swapState;
   const customPTXHandlers = usePTXCustomHandlers(manifest);
 
-  const { fromCurrency, addressFrom, toCurrency, addressTo } = useMemo(() => {
+  const { fromCurrency, addressFrom, addressTo } = useMemo(() => {
     const [, , fromCurrency, addressFrom] =
       getAccountIdFromWalletAccountId(swapState?.fromAccountId || "")?.split(":") || [];
 
@@ -208,15 +193,17 @@ const SwapWebView = ({
     const searchParams = new URLSearchParams();
 
     const swapParams = {
-      provider: swapState?.provider,
-      from: sourceCurrencyId,
-      to: targetCurrencyId,
-      amountFrom: swapState?.fromAmount,
-      loading: swapState?.loading,
       addressFrom: addressFrom,
       addressTo: addressTo,
+      amountFrom: swapState?.fromAmount,
+      from: sourceCurrencyId,
+      hasError: swapState?.error ? "true" : undefined, // append param only if error is true
+      isMaxEnabled: isMaxEnabled,
+      loading: swapState?.loading,
       networkFees: swapState?.estimatedFees,
       networkFeesCurrency: fromCurrency,
+      provider: swapState?.provider,
+      to: targetCurrencyId,
     };
 
     Object.entries(swapParams).forEach(([key, value]) => {
@@ -228,15 +215,17 @@ const SwapWebView = ({
 
     return searchParams.toString();
   }, [
-    addressFrom,
-    addressTo,
-    fromCurrency,
-    swapState?.estimatedFees,
     swapState?.fromAmount,
     swapState?.loading,
+    swapState?.estimatedFees,
     swapState?.provider,
-    targetCurrencyId,
+    swapState?.error,
+    addressFrom,
+    addressTo,
     sourceCurrencyId,
+    isMaxEnabled,
+    fromCurrency,
+    targetCurrencyId,
   ]);
 
   // return loader???
@@ -261,23 +250,6 @@ const SwapWebView = ({
       );
     }
   };
-
-  // Keep the previous UI
-  // Display only the disabled swap button
-  if (
-    isDemo1Enabled &&
-    (swapState.error ||
-      swapState.fromAmount === "0" ||
-      !(fromCurrency && addressFrom && toCurrency && addressTo))
-  ) {
-    return (
-      <Box width="100%">
-        <Button width="100%" disabled>
-          {t("sidebar.swap")}
-        </Button>
-      </Box>
-    );
-  }
 
   return (
     <>