+++ apps/web-tools/trustchain/components/JsonEditor.tsx
@@ -1,46 +1,25 @@
-import React, { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import type ace from "brace";
+import React, { useMemo } from "react";
+import CodeMirror from "@uiw/react-codemirror";
+import { json } from "@codemirror/lang-json";
 
 export function JsonEditor({
   value,
   onChange,
+  readOnly,
 }: {
   value: string;
   onChange: (value: string) => void;
+  readOnly?: boolean;
 }) {
-  const initialValueRef = useRef(value);
-  const editorRef = useRef<ace.Editor | null>(null);
+  const extensions = useMemo(() => [json()], []);
 
-  // on value changes and it's diff from editor, we force it
-  useEffect(() => {
-    const timeout = setTimeout(() => {
-      if (editorRef.current && editorRef.current.getValue() !== value) {
-        console.log("FORCE", value);
-        editorRef.current.setValue(value);
-      }
-    }, 100);
-    return () => {
-      clearTimeout(timeout);
-    };
-  }, [value]);
-
-  useEffect(() => {
-    if (typeof window === "undefined") return;
-    const ace = require("brace");
-    require("brace/mode/json");
-    require("brace/theme/github");
-    const editor = ace.edit("editor");
-    editor.setTheme("ace/theme/github");
-    editor.getSession().setMode("ace/mode/json");
-    editor.setValue(initialValueRef.current);
-    editorRef.current = editor;
-    editor.on("change", () => {
-      onChange(editor.getValue());
-    });
-    return () => {
-      editor.destroy();
-    };
-  }, [onChange]);
-
-  return <div id="editor" style={{ height: "300px" }} />;
+  return (
+    <CodeMirror
+      value={value}
+      onChange={onChange}
+      readOnly={readOnly}
+      height="300px"
+      extensions={extensions}
+    />
+  );
 }